backend/recipe-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.1
pydantic==2.4.2
python-dotenv==1.0.1
motor==3.3.1
httpx==0.27.0
prometheus-client==0.19.0 